Clean things up with thunk
