+++ app/webpack/utils/createStore.js
@@ -1,13 +1,23 @@
 import {
   createStore,
   combineReducers,
-  compose
+  compose,
+  applyMiddleware
 } from 'redux';
+import thunk from 'redux-thunk'
 
 export default (reducers) => {
   // Enable Chrome Redux DevTools extension if available
   const enableDevTools =
     window.devToolsExtension ? window.devToolsExtension() : f => f;
 
-  return compose(enableDevTools)(createStore)(combineReducers(reducers));
+  // This is awful looking.
+  return compose(
+    applyMiddleware(thunk),
+    enableDevTools
+  )(
+    createStore
+  )(
+    combineReducers(reducers)
+  );
 };
