+++ package.json
@@ -35,6 +35,7 @@
     "react-redux": "^5.0.2",
     "redux": "^3.6.0",
     "redux-actions": "^1.2.1",
+    "redux-thunk": "^2.2.0",
     "webpack": "^2.1.0-beta.25",
     "whatwg-fetch": "^1.0.0"
   }
